Moved protocol functions to respective files
